% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlo.R
\name{calculate_moving_average}
\alias{calculate_moving_average}
\title{Calculate the moving average for a time series}
\usage{
calculate_moving_average(data, window_size)
}
\arguments{
\item{data}{Matrix or Data frame.
The time series data}

\item{window_size}{Numeric integer.
The size of the moving average window.}
}
\value{
Matrix or Data frame containing the moving average values.
}
\description{
This function calculates the moving average for a time series.
}
