% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_miniconda.R
\name{conda_check}
\alias{conda_check}
\title{Check if the "transforEmotion" conda environment exists}
\usage{
conda_check()
}
\value{
A logical value indicating whether the "transforEmotion" conda
environment exists.
}
\description{
This function checks if the "transforEmotion" conda environment exists by
running the command "conda env list" and searching for the environment name
in the output.
}
