% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlo.R
\name{generate_q}
\alias{generate_q}
\title{Generate a matrix of Dynamic Error values for the DLO simulation}
\usage{
generate_q(num_steps, sigma_q)
}
\arguments{
\item{num_steps}{Numeric integer.
The number of time steps used in the simulation.}

\item{sigma_q}{Numeric. 
Standard deviation of the Dynamic Error/}
}
\value{
A (num_steps X 3) matrix of Dynamic Error values for neutral, negative and positive emotion latent score.
}
\description{
This function generates a matrix of Dynamic Error values (q) for the DLO simulation.
}
