% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo_ipip_extraversion.R
\docType{data}
\name{neo_ipip_extraversion}
\alias{neo_ipip_extraversion}
\title{NEO-PI-R IPIP Extraversion Item Descriptions}
\format{
A list (length = 6)
}
\usage{
data(neo_ipip_extraversion)
}
\description{
A list (length = 6) of the NEO-PI-R IPIP item descriptions
(https://ipip.ori.org/newNEOFacetsKey.htm). Each vector within
the 6 list elements contains the item descriptions for the
respective Extraversion facets -- friendliness, gregariousness,
assertiveness, activity_level, excitement_seeking, and
cheerfulness
}
\examples{
data("neo_ipip_extraversion")

}
\keyword{datasets}
