% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_emotions.R
\name{print.emotion_evaluation}
\alias{print.emotion_evaluation}
\title{Print method for emotion evaluation results}
\usage{
\method{print}{emotion_evaluation}(x, ...)
}
\arguments{
\item{x}{An emotion_evaluation object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for emotion evaluation results
}
