% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_setup.R
\name{python_setup}
\alias{python_setup}
\title{Setup Python}
\usage{
python_setup(path_to_python = NULL)
}
\arguments{
\item{path_to_python}{Character.
Path to specify where "python.exe" is located on your computer.
Defaults to \code{NULL}, which will use \code{\link[reticulate]{py_available}}
to find available Python or Anaconda}
}
\value{
Returns a message either specifying help or enables Python
}
\description{
Sets up Python
}
\examples{
python_setup()

}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
