% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gpu_modules.R
\name{setup_gpu_modules}
\alias{setup_gpu_modules}
\title{Install GPU Python Modules}
\usage{
setup_gpu_modules()
}
\description{
Adds \code{bitsandbytes}, used for 4-bit quantization of the EVA-CLIP-8B
vision model, to the Python environment. Equivalent to
\code{setup_modules(extras = "gpu", download_models = FALSE)}.
}
\details{
The CUDA build of PyTorch does not need this function: it is selected
automatically when an NVIDIA GPU is detected (see \code{\link{setup_modules}}).
\code{image_scores()} and \code{video_scores()} add \code{bitsandbytes}
themselves when EVA-CLIP-8B is used on a GPU.
}
\note{
This function requires NVIDIA GPU and drivers to be properly installed.
}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
