% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_words.R
\docType{data}
\name{stop_words}
\alias{stop_words}
\title{Stop Words from the \emph{tm} Package}
\format{
A vector (length = 174)
}
\usage{
data(stop_words)
}
\description{
174 English stop words in the \emph{tm} package
}
\examples{
data("stop_words")

}
\keyword{datasets}
