% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_emotions.R
\name{summary.emotion_evaluation}
\alias{summary.emotion_evaluation}
\title{Summary method for emotion evaluation results}
\usage{
\method{summary}{emotion_evaluation}(object, ...)
}
\arguments{
\item{object}{An emotion_evaluation object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for emotion evaluation results
}
