% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlo.R
\name{plot_sim_emotions}
\alias{plot_sim_emotions}
\title{Plot the latent or the observable emotion scores.}
\usage{
plot_sim_emotions(df, mode = "latent", title = " ")
}
\arguments{
\item{df:}{Data frame.
The data frame containing the latent and observable variables created by the \code{simulate_video} function.}

\item{mode:}{Character.
The mode of the plot. Can be either 'latent', 'positive' or 'negative'.}
}
\value{
A plot of the latent or the observable emotion scores.
}
\description{
Function to plot the latent or the observable emotion scores.
}
\examples{
plot_sim_emotions(sim_results, 'latent')
Updated: 26.10.2023.
}
