% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gpu_modules.R
\name{setup_gpu_modules}
\alias{setup_gpu_modules}
\title{Install GPU Python Modules}
\usage{
setup_gpu_modules()
}
\description{
Installs GPU modules for the {transforEmotion} conda environment
}
\details{
Installs modules for miniconda using \code{\link[reticulate]{conda_install}}
}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
