% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_miniconda.R
\name{setup_miniconda}
\alias{setup_miniconda}
\title{Install Miniconda}
\usage{
setup_miniconda()
}
\description{
Installs miniconda
}
\details{
Installs miniconda using \code{\link[reticulate]{install_miniconda}}
}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
