% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_modules.R
\name{setup_modules}
\alias{setup_modules}
\title{Install Necessary Python Modules}
\usage{
setup_modules()
}
\description{
Installs modules for the \{transforEmotion\} conda environment.
}
\details{
This function installs the required Python modules for the \{transforEmotion\} conda environment using \code{\link[reticulate]{conda_install}}. Ensure that miniconda is installed and properly configured before running this function.
}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
