% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_modules.R
\name{setup_modules}
\alias{setup_modules}
\title{Install Necessary Python Modules}
\usage{
setup_modules()
}
\description{
Installs modules for the \{transforEmotion\} conda environment
}
\details{
Installs modules for miniconda using \code{\link[reticulate]{conda_install}}
}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
