% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_modules.R
\name{setup_modules}
\alias{setup_modules}
\title{Install Necessary Python Modules}
\usage{
setup_modules()
}
\description{
Installs modules to compute \code{\link[transforEmotion]{transformer_scores}}. These include
\itemize{

\item{pytorch}

\item{torchvison}

\item{torchaudio}

\item{tensorflow}

\item{transformers}

\item{pytube}

\item{pytx}

\item{pandas}

\item{opencv-python}

\item{accelerate}

}
}
\details{
Installs modules for miniconda using \code{\link[reticulate]{conda_install}}
}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
